Replace assert(false) with fatalError()
